+++ RunningStatistics/IRunningStatistic.cs
@@ -4,7 +4,7 @@ namespace RunningStatistics;
 
 public interface IRunningStatistic<TObs>
 {
-    public long Count { get; }
+    public nint Count { get; }
 
     public void Fit(IEnumerable<TObs> values);
