+++ RunningStatistics/Moments.cs
@@ -30,7 +30,7 @@ public class Moments : IRunningStatistic<double>
     }
     
 
-    public long Count { get; private set; }
+    public nint Count { get; private set; }
     public double Mean => Count == 0 ? double.NaN : _mean;
 
     /// <summary>
